Refactor generateDID in helper.js
