+++ packages/transmute-did/src/__tests__/tst-decentralized-ledger-interface.spec.js
@@ -1,29 +1,5 @@
 const dli = require('../tst-decentralized-ledger-interface.js');
-const { misc, msg, pgp } = require('../index.js');
-
-// Helper function
-const generateDID = async () => {
-    // Generate keys
-    const pgpKeypair = await pgp.generateOpenPGPArmoredKeypair({
-        name: "bob",
-        passphrase: "yolo"
-    });
-    const libSodiumSigningKeypair = await msg.generateCryptoSignKeypair();
-    const libSodiumEncryptionKeypair = await msg.generateCryptoBoxKeypair();
-    const primaryPublicKey = libSodiumSigningKeypair.publicKey;
-    // Generate DID
-    const did = misc.publicKeyToTransmuteDID({publicKey: primaryPublicKey});
-    // Generate DID document
-    const didDocumentArgs = {
-        primaryPublicKey,
-        pgpPublicKey: pgpKeypair.publicKey,
-        libSodiumPublicSigningKey: libSodiumSigningKeypair.publicKey,
-        libSodiumPublicEncryptionKey: libSodiumEncryptionKeypair.publicKey
-    };
-    const didDocument = await misc.publicKeysToDIDDocument(didDocumentArgs);
-
-    return {did, didDocument};
-}
+const { generateDID } = require('../helper.js')
 
 let did1, didDocument1;
 let did2;
