+++ packages/transmute-did/src/tst-decentralized-ledger-interface.js
@@ -1,4 +1,5 @@
 // TODO: eslint error
+// TODO: Switch to IPFS instead of mocked decentralizedLedger
 const decentralizedLedger = {};
 
 const create = (did, didDocument) => {
